update R dependency search; allow bioc
This allows bioconductor packages to be installed and cleans up some of
the steps from the workflows to make them a bit more readable.

A new function is added to bin/dependencies.R called use_bioc_repos() that will
set "repos" option to the ones used by BioConductor and returns a function that
resets it to the user-set repository options.

BiocManager has been added to the default packages added.